+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 (Add entries below in the order of server, console, cli, docs, others)
 
+- server: set `tracecontext` and `userInfo` for DML actions on Postgres sources
 - cli: add progress bar for `migrate apply` command (#4795)
 
 ## v2.0.8